# Reduce number of allocations.