--- a/.travis.yml
+++ b/.travis.yml
@@ -1,10 +1,14 @@
+sudo: false
 language: go
 
 go:
-  - 1.2
-  - 1.3
-  - 1.4
-  - 1.5
+  - 1.6
+  - 1.7
+  - tip
+
+matrix:
+  allow_failures:
+    - go: tip
 
 install:
   - mkdir -p $HOME/gopath/src/gopkg.in