--- a/Makefile
+++ b/Makefile
@@ -1,3 +1,4 @@
 all:
-	go test ./... -test.v -test.cpu=1,2,4
-	go test ./... -test.v -test.short -test.race
+	go test ./...
+	go test ./... -short -race
+	go vet