--- a/consumer_test.go
+++ b/consumer_test.go
@@ -175,29 +175,6 @@ func TestJQuerySourceMap(t *testing.T) {
 	}
 }
 
-func BenchmarkParse(b *testing.B) {
-	for i := 0; i < b.N; i++ {
-		_, err := sourcemap.Parse(jqSourceMapURL, jqSourceMapBytes)
-		if err != nil {
-			b.Fatal(err)
-		}
-	}
-}
-
-func BenchmarkSource(b *testing.B) {
-	smap, err := sourcemap.Parse(jqSourceMapURL, jqSourceMapBytes)
-	if err != nil {
-		b.Fatal(err)
-	}
-	b.ResetTimer()
-
-	for i := 0; i < b.N; i++ {
-		for j := 0; j < 10; j++ {
-			smap.Source(j, 100*j)
-		}
-	}
-}
-
 // This is a test mapping which maps functions from two different files
 // (one.js and two.js) to a minified generated source.
 //