```
benchmark            old ns/op     new ns/op     delta
BenchmarkParse-4     10292833      5035375       -51.08%

benchmark            old allocs     new allocs     delta
BenchmarkParse-4     99529          2312           -97.68%

benchmark            old bytes     new bytes     delta
BenchmarkParse-4     5953245       4397716       -26.13%
```